refactor: extract active cleanup trait
